requirements.txt
google-adk
litellm
pydantic
typing_extensions
aiosqlite
yfinance